@@ -1,8 +1,9 @@
 {
   "name": "vigi",
-  "version": "0.0.0",
-  "description": "Browser for dalet, text, gemtext (gemini protocol included) with native support of txtdot proxy and local txtdot engines",
+  "version": "0.1.0",
+  "description": "",
   "type": "module",
+  "private": true,
   "scripts": {
     "dev": "vite dev",
     "build": "vite build",
@@ -13,22 +14,27 @@
   },
   "license": "MIT",
   "dependencies": {
-    "@tauri-apps/api": "^1",
-    "@txtdot/dalet": "1.0.0-pre3"
+    "@tailwindcss/vite": "^4.1.5",
+    "@tauri-apps/api": "^2.5.0",
+    "@tauri-apps/plugin-opener": "^2.2.6",
+    "@txtdot/dalet": "1.0.0-pre10",
+    "svelte-dnd-action": "^0.9.61",
+    "svelte-highlight": "^7.8.3",
+    "tailwindcss": "^4.1.5"
   },
   "devDependencies": {
-    "@iconify/svelte": "^4.0.2",
-    "@sveltejs/adapter-static": "^3.0.1",
-    "@sveltejs/kit": "^2.0.0",
-    "@sveltejs/vite-plugin-svelte": "^3.0.0",
-    "@tauri-apps/cli": "^1",
-    "autoprefixer": "^10.4.19",
-    "postcss": "^8.4.40",
-    "svelte": "^4.2.7",
-    "svelte-check": "^3.6.0",
-    "tailwindcss": "^3.4.7",
-    "tslib": "^2.4.1",
-    "typescript": "^5.0.0",
-    "vite": "^5.0.3"
+    "@sveltejs/adapter-static": "^3.0.8",
+    "@sveltejs/kit": "^2.20.8",
+    "@sveltejs/vite-plugin-svelte": "^5.0.3",
+    "@tauri-apps/cli": "^2.5.0",
+    "svelte": "^5.28.2",
+    "svelte-check": "^4.1.7",
+    "typescript": "~5.6.3",
+    "vite": "^6.3.4"
+  },
+  "pnpm": {
+    "onlyBuiltDependencies": [
+      "esbuild"
+    ]
   }
 }
