@@ -1,17 +1,20 @@
 <svg
   xmlns="http://www.w3.org/2000/svg"
-  width="1em"
-  height="1em"
+  width="24"
+  height="24"
   viewBox="0 0 24 24"
   {...$$props}
 >
-  <path
+  <g
     fill="none"
     stroke="currentColor"
     stroke-linecap="round"
     stroke-linejoin="round"
-    stroke-width="1.5"
-    d="m15.167 1l.598 1.118c.404.755.606 1.133.472 1.295c-.133.162-.573.031-1.454-.23A9.8 9.8 0 0 0 12 2.78c-5.247 0-9.5 4.128-9.5 9.22a8.97 8.97 0 0 0 1.27 4.61M8.834 23l-.598-1.118c-.404-.756-.606-1.134-.472-1.295c.133-.162.573-.032 1.454.23c.88.261 1.815.402 2.783.402c5.247 0 9.5-4.128 9.5-9.22a8.97 8.97 0 0 0-1.27-4.609"
-    color="currentColor"
-  />
+    stroke-width="2"
+  >
+    <path
+      d="M19.933 13.041a8 8 0 1 1-9.925-8.788c3.899-1 7.935 1.007 9.425 4.747"
+    />
+    <path d="M20 4v5h-5" />
+  </g>
 </svg>
