@@ -1,16 +1,38 @@
 import type { Tag } from "@txtdot/dalet";
 
-export interface VigiState {
-  current_tab_index: number;
-  tabs: StateTab[];
-  favorites_tabs: StateTab[];
+export type TabType = "RENDER" | "BROWSER";
 
-  top_bar_input: string;
-  current_data: Tag[];
+export interface DrovaError {
+  body?: string;
+  message: string;
 }
 
-export interface StateTab {
-  title: string;
-  url: string;
+export interface VigiState {
+  current_tab: number;
+  tab_counter: number;
+  tabs: SiteTab[];
+}
+
+export interface SiteTab {
   id: number;
+  current_link: number;
+  links: TabLink[];
+}
+
+export interface TabLink {
+  title?: string;
+  body?: Tag[];
+
+  ty: TabType;
+  uri: string;
+
+  error?: DrovaError;
+  loading?: true;
+}
+
+export interface TemporalState {
+  top_bar_open: boolean;
+  first_load: boolean;
+  sidebar_scroll: number;
+  sidebar_open: boolean;
 }
