@@ -1,5 +1,5 @@
 // Tauri doesn't have a Node.js server to do proper SSR
 // so we will use adapter-static to prerender the app (SSG)
-// See: https://beta.tauri.app/start/frontend/sveltekit/ for more info
+// See: https://v2.tauri.app/start/frontend/sveltekit/ for more info
 export const prerender = true;
 export const ssr = false;
